+++ vistrails/core/vistrail/abstraction.py
@@ -139,7 +139,7 @@ class Abstraction(DBAbstraction, Module):
             return
         latest_version = desc.module.vistrail.get_latest_version()
         self._is_latest_version = \
-            (long(latest_version) == long(self.internal_version))
+            (latest_version == self.internal_version)
 
     def is_latest_version(self):
         return self._is_latest_version
